Fix comprehensive test CI failures
