+++ crates/llama-quant-ffi/build.rs
@@ -322,6 +322,7 @@ fn emit_system_links(
         println!("cargo:rustc-link-lib=dylib=m");
         println!("cargo:rustc-link-lib=dylib=dl");
         println!("cargo:rustc-link-lib=dylib=pthread");
+        link_linux_openmp(cmake_cache);
         if has_cuda {
             link_linux_cuda_libs(cmake_cache);
         }
@@ -331,6 +332,35 @@ fn emit_system_links(
         if has_vulkan {
             println!("cargo:rustc-link-lib=dylib=vulkan");
         }
+    }
+}
+
+fn link_linux_openmp(cmake_cache: &std::path::Path) {
+    let Ok(cache) = std::fs::read_to_string(cmake_cache) else {
+        return;
+    };
+    if cmake_cache_value(&cache, "GGML_OPENMP_ENABLED").as_deref() != Some("ON") {
+        return;
+    }
+
+    let mut libraries = Vec::new();
+    for cache_key in ["OpenMP_C_LIB_NAMES", "OpenMP_CXX_LIB_NAMES"] {
+        if let Some(names) = cmake_cache_value(&cache, cache_key) {
+            for library in names
+                .split(';')
+                .filter(|library| !library.is_empty() && *library != "pthread")
+            {
+                if !libraries.iter().any(|existing| existing == library) {
+                    libraries.push(library.to_string());
+                }
+            }
+        }
+    }
+    if libraries.is_empty() {
+        libraries.push("gomp".to_string());
+    }
+    for library in libraries {
+        link_linux_lib_from_cache(cmake_cache, &format!("OpenMP_{library}_LIBRARY"), &library);
     }
 }
