+++ crates/mesh-llm-host-runtime/src/mesh/tests/mod.rs
@@ -3,6 +3,8 @@ use crate::mesh::heartbeat::HeartbeatFailurePolicy;
 use crate::mesh::node::LocalRequestMetricsSampler;
 use mesh_llm_types::mesh::{DEMAND_TTL_SECS, merge_demand};
 
+mod control_listener;
+
 include!("connections.rs");
 include!("admission/helpers.rs");
 include!("protocol_compat.rs");
